fix: move conftest up one so submodules see it
